Moved and renamed view models
