--- GIFrameworkMap.Data/Data/Models/ViewModels/ApplicationUserRoleViewModel.cs
@@ -1,9 +1,0 @@
-﻿namespace GIFrameworkMaps.Data.Models.ViewModels
-{
-	public class ApplicationUserRoleViewModel
-    {
-        public string? UserId { get; } 
-        public int RoleId { get; }
-        public string? RoleName { get; }
-    }
-}
