--- GIFrameworkMap.Data/Data/Models/ViewModels/BasemapViewModel.cs
@@ -1,17 +1,0 @@
-﻿namespace GIFrameworkMaps.Data.Models.ViewModels
-{
-	public class BasemapViewModel
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-        public bool IsDefault { get; set; }
-        public string? PreviewImageURL { get; set; }
-        public Bound? Bound { get; set; }
-        public int MaxZoom { get; set; }
-        public int MinZoom { get; set; }
-        public int DefaultOpacity { get; set; }
-        public int DefaultSaturation { get; set; }
-        public int SortOrder { get; set; }
-        public LayerSource? LayerSource { get; set; }
-    }
-}
