+++ GIFrameworkMap.Data/ViewModels/ApplicationUserRoleViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace GIFrameworkMaps.Data.ViewModels
+{
+	public class ApplicationUserRoleViewModel
+	{
+		public string? UserId { get; }
+		public int RoleId { get; }
+		public string? RoleName { get; }
+	}
+}
